Add more robust socket detection
My user env is not setup right, and the scripts aren't being fed the
right variables. Sometimges that's an issue. Luckily, the sway ipc
socket location is predictable, so we can fix it.

This is not compatible with having multiple users running multiple sway
sessions.  No idea why you'd do that.